% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_v28_cmshcc_hierarchy.R
\name{apply_v28_cmshcc_hierarchy}
\alias{apply_v28_cmshcc_hierarchy}
\title{Apply v28 CMS-HCC hierarchy to binary data.table}
\usage{
apply_v28_cmshcc_hierarchy(curr_dt)
}
\arguments{
\item{curr_dt}{(data.table, also accepts data.frame or tibble) The input binary data.table, with rows corresponding to
beneficiaries (with numeric row ID) and columns corresponding to HCCs (in
format "hcc*")}
}
\value{
curr_dt (data.table) The input data.table with hierarchy rules applied
}
\description{
This function applies the v28 hierachy to the input binary data.table
}
