% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_by_row_ids.R
\name{censor_by_row_ids}
\alias{censor_by_row_ids}
\title{Censor a specified data.table at specific row_ids.}
\usage{
censor_by_row_ids(curr_dt, all_censoring_row_id_tbl, t)
}
\arguments{
\item{curr_dt}{(data.table) The data.table to upcode. This is a binary data.table with
rows corresponding to people and columns corresponding to 115 HCCs (from v28 of CMS-HCC) and one column for
person_ids. If it has not been modified already, this is generated by simulate_baseline_v28_data}

\item{all_censoring_row_id_tbl}{(tibble) All row ids to right censor across all time points}

\item{t}{(int) The current time point, must be greater than or equal to 1}
}
\description{
This function changes the values at \verb{[hcc, row_id]} for each of the row_ids
from 0 or 1 to 0 in place.
}
