% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_curr_dt.R
\name{check_curr_dt}
\alias{check_curr_dt}
\title{Verifies whether curr.dt is a data.table and has the expected columns. Expected columns
should be "person_id" followed by the HCCs in version 28 of the CMS-HCC risk adjustment formula.
As a note, you can access these HCCs with \code{v28_hccs} (they are a global variable in this package).}
\usage{
check_curr_dt(curr_dt)
}
\arguments{
\item{curr_dt}{(data.table) The data.table to verify. This should be a binary data.table with
rows corresponding to people and columns corresponding to 115 HCCs (from v28 of CMS-HCC) and one column for
person_ids. If it has not been modified already, this is generated by simulate_baseline_v28_data}
}
\description{
Verifies whether curr.dt is a data.table and has the expected columns. Expected columns
should be "person_id" followed by the HCCs in version 28 of the CMS-HCC risk adjustment formula.
As a note, you can access these HCCs with \code{v28_hccs} (they are a global variable in this package).
}
