% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_upcoding_spec_df.R
\name{check_upcoding_spec_df}
\alias{check_upcoding_spec_df}
\title{Checks whether upcoding specification dataframes are formatted correctly}
\usage{
check_upcoding_spec_df(upcoding_spec_df)
}
\arguments{
\item{upcoding_spec_df}{(tibble or dataframe) Specifies which HCCs to upcode and how. This
is expected to have columns 'hcc' (containing any of the v28 HCCs), 'approach'
(either 'any' or 'lower severity' are allowed), and 'upcoding_prop' (must be a proportion
greater than 0 and less than 1)}
}
\description{
Verifies whether upcoding_spec_df is a tibble or data frame,column names,
whether hccs are valid v28 hccs, whether approaches are one of valid options,
and if upcoding_proportions are between 0 and 1.
}
