% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cms_hcc_v28_reference_hcc_count_dt}
\alias{cms_hcc_v28_reference_hcc_count_dt}
\title{Co-occuring v28 HCCs based on summarized AllOfUs self-reported survey data with missing HCCs imputed}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 219 rows and 2 columns.
}
\source{
Self-reported survey data from All Of Us v7 (https://www.researchallofus.org/) with missing HCCs imputed. The
code to construct this from AllOfUs counts (a summarized version of which is available as 'coocurring_v28_hcc_counts_allofus')
is available in R/make_reference_dataset.R in the function 'make_reference_dataset()'.
}
\usage{
cms_hcc_v28_reference_hcc_count_dt
}
\description{
A tibble with two columns, 'hcc_set' and 'num_respondents'. Each row of 'hcc_set' corresponds to a
unique combination of co-occurring HCCs (from version 28 of the CMS-HCC risk adjustment formula) as a single character.
Missing HCCs were imputed and randomly inserted into summary AllOfUs data; see 'source' below for details.
All 'num_respondents' values are set to counts > 20, which aligns with All of Us policies; however, note that values set for imputed
HCCs (i.e. HCCs not included in \code{coocurring_v28_hcc_counts_all_of_us}) are not real.
}
\keyword{datasets}
