% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooccurring_v28_hcc_counts_allofus}
\alias{cooccurring_v28_hcc_counts_allofus}
\title{Co-occuring v28 HCCs based on summarized AllOfUs self-reported survey data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 134 rows and 2 columns.
}
\source{
Self-reported survey data from All Of Us v7 (https://www.researchallofus.org/).
Lower level data are not allowed to be distributed outside the AllOfUs Research Workbench, but we provide the code we used
to generate these data in the manuscript associated with this package (TODO add link here).
}
\usage{
cooccurring_v28_hcc_counts_allofus
}
\description{
A tibble with two columns, 'hcc_set' and 'num_respondents'. Each row of 'hcc_set' corresponds to a
unique combination of co-occurring HCCs (from version 28 of the CMS-HCC risk adjustment formula) as a single character.
'num_respondents' is a count > 21 of how many respondents were found to have a given
co-occurring set of v28 HCCs. This complies with All of Us policies because it both excludes any counts of 20 or fewer and a number of counts greater than 20.
}
\keyword{datasets}
