% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_row_ids_to_censor.R
\name{get_all_row_ids_to_censor}
\alias{get_all_row_ids_to_censor}
\title{Get overall list of row IDs to censor for an individual HCC}
\usage{
get_all_row_ids_to_censor(curr_dt, num_timepoints, censoring_prop)
}
\arguments{
\item{curr_dt}{(data.table) Binary data.table with rows corresponding to people and
columns corresponding to 115 HCCs (from v28 of CMS-HCC) and one column for
person_ids.}

\item{num_timepoints}{(int) The number of timepoints across which to censor sets of beneficiaries}

\item{censoring_prop}{(double) The proportion of beneficiaries to
censor overall (across all timepoints). Must be greater than 0 and less than 1, exclusive.}
}
\value{
A tibble of row IDs to censor at each time point
}
\description{
This samples (without replacement) a list of row indexes to censor from the
overall dataset, so that the user-specified upcoding proportion and approach
occurs.
}
