% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_row_ids_to_upcode.R
\name{get_all_row_ids_to_upcode}
\alias{get_all_row_ids_to_upcode}
\title{Get all row IDs to upcode for all HCCs specified in upcoding_spec_df}
\usage{
get_all_row_ids_to_upcode(upcoding_spec_df, curr_dt, num_timepoints)
}
\arguments{
\item{upcoding_spec_df}{(tibble or dataframe) Tibble or data frame specifying which HCCs to upcode and how. This
is expected to have columns 'hcc' (containing any of the v28 HCCs), 'approach'
(either 'any' or 'lower severity' are allowed), and 'upcoding_prop' (must be a proportion
greater than 0 and less than 1)}

\item{curr_dt}{(data.table) Binary data.table with rows corresponding to people and
columns corresponding to 115 HCCs (from v28 of CMS-HCC) and one column for
person_ids. If it has not been modified already, this is generated by simulate_baseline_v28_data}

\item{num_timepoints}{(int) The number of timepoints across which to censor sets of beneficiaries}
}
\value{
A tibble of row IDs to upcode for each HCC in upcoding_spec_df across the number of timepoints specified
}
\description{
For each HCC, this samples (without replacement) a list of row indexes to upcode from the
overall dataset, so that the user-specified upcoding proportion and approach
occurs.
}
