% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcc_hierarchy_list}
\alias{hcc_hierarchy_list}
\title{Hierarchy of hierarchical condition categories in CMS-HCC risk adjusment version 28}
\format{
\subsection{\code{hcc_hierarchy_list}}{

A nested list, where the name of the nested list is the head of the branch
}
}
\source{
https://www.cms.gov/medicare/payment/medicare-advantage-rates-statistics/risk-adjustment/2025-model-software/icd-10-mappings
}
\usage{
hcc_hierarchy_list
}
\description{
A nested list describing the hierarchical structure of version 28 of
the CMS-HCC risk adjustment algorithm.
}
\keyword{datasets}
