% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reference_dataset.R
\name{make_reference_dataset}
\alias{make_reference_dataset}
\title{Make reference data set}
\usage{
make_reference_dataset(
  coocurring_hcc_counts = cooccurring_v28_hcc_counts_allofus,
  missing_hcc_respondent_count = 21,
  randomly_insert_missing_hccs_into_cooccurring_sets = TRUE,
  missing_hcc_insertion_proportion = 0.25,
  curr_seed = 123
)
}
\arguments{
\item{coocurring_hcc_counts}{(tibble, Default: \code{cooccurring_v28_hcc_counts_allofus}). What
reference set of co-occurring HCC sets to use for building the reference data. This tibble
is expected to have two columns (\code{hcc_set} and \code{num_respondents}) and (after inserting
missing HCCs, if needed) will be used to sample rows for the baseline data. \code{hcc_set}
corresponds to co-occurring HCCs to code as being present in baseline data, and \code{num_respondents}
acts as a weight for sampling (so HCC sets with higher \code{num_respondent} values will be sampled
more in creating baseline data).}

\item{missing_hcc_respondent_count}{(int, Default: 21) The respondent count to use for HCCs
that are missing from co-occurring_hcc_counts. By default this value is 21, consistent with
AllOfUs' minimum value requirement for data dissemination.}

\item{randomly_insert_missing_hccs_into_cooccurring_sets}{(boolean, Default: TRUE). Whether
to randomly insert HCCs not included in cooccurring_hcc_counts into existing HCC sets or not. The
reason a user might want to do this is to not disproportionately penalize missing HCCs in
baseline sampling in the future (assuming there are not many missing HCCs relative to the
overall number of co-occurring HCC sets).}

\item{missing_hcc_insertion_proportion}{(double, Default: 0.25) What proportion of the time
to insert HCCs not included in cooccurring_hcc_counts into existing HCC sets when building the
reference dataset to sample from. Must be greater than 0 and less than 1, exclusive. Each
HCC is inserted separately into missing_hcc_insertion_proportion times the number of existing reference sets.}

\item{curr_seed}{(int, Default: 123) Seed for reproducibility}
}
\description{
This is an optional helper function for simulate_baseline_v28_hcc_dt, in case users
don't want to use the default value for reference_cooccurring_hcc_count (which is
cms_hcc_v28_reference_hcc_count_dt).
}
