% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_v28_cmshcc_hierarchy.R
\name{set_to_zero}
\alias{set_to_zero}
\title{Helper function to set lower hierarchy columns to 0 if condition_col is 1}
\usage{
set_to_zero(curr_dt, cols, condition_col)
}
\arguments{
\item{curr_dt}{(data.table) The data.table to modify}

\item{cols}{(vector of strings) The columns to set to zero if condition_col is 1}

\item{condition_col}{(string) The column to check about being 1 or 0}
}
\description{
This function modifies a data.table object in place
}
