% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_baseline_v28_hcc_dt.R
\name{simulate_baseline_v28_hcc_dt}
\alias{simulate_baseline_v28_hcc_dt}
\title{Simulate baseline HCC values using reference hcc sets and CMS-HCC v28
hierarchy}
\usage{
simulate_baseline_v28_hcc_dt(
  num_people,
  out_dir,
  reference_cooccurring_hcc_count = cms_hcc_v28_reference_hcc_count_dt,
  apply_v28_hierarchy = FALSE,
  out_file_prefix = "baseline_data",
  curr_seed = 123
)
}
\arguments{
\item{num_people}{(int) The number of people (e.g. rows) to simulate, represented
as an integer.}

\item{out_dir}{(string) Directory to write files}

\item{reference_cooccurring_hcc_count}{(data.table, also accepts dataframe or tibble Default: \code{cms_hcc_v28_reference_hcc_count_dt})}

\item{apply_v28_hierarchy}{(boolean, Default: FALSE) Whether to apply the v28 hierarchy to the data after upcoding}

\item{out_file_prefix}{(string, default: \code{"baseline_data"}) Prefix used for output
files. The dataset will be written as \verb{[out_file_prefix].csv}, and timestamps
as \verb{[out_file_prefix]_times.csv}.}

\item{curr_seed}{(int, Default: 123) Seed for reproducibility}
}
\description{
This function simulates a binary data.table of 115 HCCs corresponding to
those used in version 28 (v28) of the CMS-HCC risk adjustment payment model.
It samples from a reference dataset of co-occurring HCCs sets with counts (e.g. weights);
see cms_hcc_v28_reference_hcc_count_dt (the default reference dataset) or the documentation of
make_reference_dataset for more details about this. If specified, the function
also applies the CMS-HCC v28 hierarchy to columns by row.
}
