% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_row_ids_into_timepoints.R
\name{split_row_ids_into_timepoints}
\alias{split_row_ids_into_timepoints}
\title{Splits the overall row IDs to upcode into multinomially distributed random
group sizes, where the number of groups is specified by the user-supplied
number of timepoints.}
\usage{
split_row_ids_into_timepoints(row_ids, num_timepoints)
}
\arguments{
\item{row_ids}{(vector or list) One set of row IDs to split into groups. This should be the
output of get_all_row_ids_to_upcode.}

\item{num_timepoints}{(int) The number of timepoints to use.}
}
\value{
A list of length num_timepoints with the row_ids randomly distributed across each element
}
\description{
This splits the overall set of row IDs to upcode into num_timepoints number
of groups. The group sizes are multinomially distributed with equal
probability per group, but this can be varied by the user if you prefer
alternate groupings.
}
