% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undercode_baseline_data.R
\name{undercode_dt}
\alias{undercode_dt}
\title{Undercodes (e.g. sets to zero) a specified proportion of existing diagnoses, where diagnoses are
selected randomly.}
\usage{
undercode_dt(
  curr_dt,
  undercoding_prop,
  out_dir,
  out_file_prefix = "undercoded_data",
  curr_seed = 123
)
}
\arguments{
\item{curr_dt}{(data.table; also accepts data.frame or tibble) The data.table to undercode}

\item{undercoding_prop}{(double) The proportion of all diagnoses to undercode. Must be greater
than zero and less than 1, exclusive.}

\item{out_dir}{(string) The directory to which to write output}

\item{out_file_prefix}{(string, Default: "undercoded_data") Prefix of output for undercoded
files. Note-- by default the undercoding proportion will also be added to the file
name as a suffix (e.g. undercoding_prop = 0.25 will be represented as "0pt25")}

\item{curr_seed}{(int, Default: 123) Seed for reproducibility}
}
\description{
Undercodes (e.g. sets to zero) a specified proportion of existing diagnoses, where diagnoses are
selected randomly.
}
