% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upcode_hcc_by_row_ids.R
\name{upcode_hcc_by_row_ids}
\alias{upcode_hcc_by_row_ids}
\title{Upcode a specified data.table at specific row_ids.}
\usage{
upcode_hcc_by_row_ids(curr_dt, all_upcoding_row_id_tbl, h, t)
}
\arguments{
\item{curr_dt}{(data.table; also accepts data.frame or tibble) The data.table to upcode. This is a binary data.table with
rows corresponding to people and columns corresponding to 115 HCCs (from v28 of CMS-HCC) and one column for
person_ids. If it has not been modified already, this is generated by simulate_baseline_v28_data}

\item{all_upcoding_row_id_tbl}{(tibble) Tibble of all row ids upcode across all time points}

\item{h}{(int) the hcc (e.g. column) to upcode}

\item{t}{(int) the specified time point}
}
\description{
This function changes the values at \verb{[hcc, row_id]} for each of the row_ids
from 0 to 1 in place.
}
