% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upcoding-package.R
\docType{package}
\name{upcoding-package}
\alias{upcoding}
\alias{upcoding-package}
\title{upcoding: A Package to Simulate Medicare Upcoding and Undercoding}
\description{
This package enables users to: (1) Simulate baseline data realistic to older Americans' co-occurring health conditions at scale; (2) Upcode specific diagnoses in baseline data to a user-specified degree, including altering right censoring; (3) Undercode all baseline data.
}
\author{
\strong{Maintainer}: Oana Enache \email{oenache@stanford.edu}

}
\keyword{internal}
