% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v28_hccs}
\alias{v28_hccs}
\title{Hierarchical condition categories included in CMS-HCC risk adjustment version 28}
\format{
\subsection{\code{v28_hccs}}{

A vector with 115 elements, each beginning with 'hcc' and ending with a number
}
}
\source{
https://www.cms.gov/medicare/payment/medicare-advantage-rates-statistics/risk-adjustment/2025-model-software/icd-10-mappings
}
\usage{
v28_hccs
}
\description{
A list of the hierarchical condition categories (e.g. diagnoses)
in version 28 of the CMS-HCC risk adjustment algorithm.
}
\keyword{datasets}
